feat: add Metactical Settings document with validation for default USAePay accounts
