+++ metactical/metactical/doctype/metactical_settings/metactical_settings.js
@@ -1,0 +1,8 @@
+// Copyright (c) 2025, Techlift Technologies and contributors
+// For license information, please see license.txt
+
+frappe.ui.form.on('Metactical Settings', {
+	// refresh: function(frm) {
+
+	// }
+});
