+++ metactical/metactical/doctype/metactical_settings/test_metactical_settings.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2025, Techlift Technologies and Contributors
+# See license.txt
+
+# import frappe
+from frappe.tests.utils import FrappeTestCase
+
+
+class TestMetacticalSettings(FrappeTestCase):
+	pass
