+++ metactical/metactical/doctype/website_deduct_qty/website_deduct_qty.py
@@ -4,13 +4,5 @@
 import frappe
 from frappe.model.document import Document
 
-class MetacticalSettings(Document):
-	def validate(self):
-		has_default = 0
-		for account in self.usaepay_accounts:
-			if account.is_default:
-				has_default += 1
-
-		if has_default > 1:
-			frappe.throw("Only one account can be default")
-		
+class WebsiteDeductQty(Document):
+	pass
